Use dynamic template from Sendgrid
This allows for using Sendgrid's email designer to create a template.
The template is populated with the data we are getting from Jellyfin. If
we need to send more data in the future or have different emails for
different types of content added we can do so by swapping the template.