Allow for multiple receiver emails
This commit allows for multiple receiver emails to be specified in the
SENDGRID_RECEIVER_EMAIL environment variable. The emails should be
separated by commas. The emails are then split by commas and mapped to
an array of objects with the email key and value.

SENDGRID_RECEIVER_EMAIL=email1@mail.com, email2@mail.com