Revise route names for the last time
I really hope this is the last time I need to change this. I thought
about it and while it would be cool to expand I think it is best I keep
the scope of this project to just emails (hence the name). I will be
renaming the routes to match the notification type that is sent from the
Jellyfin server.

Future routes would include things like "playback-started" or
"playback-stopped".