Read emails from the request body
This will allow the user to send emails to multiple recipients by adding
a new field to the request body handlebars. This takes precedence over
any environment variable that might be set. I like this approach better
as it allows the user to send emails to different recipients for
different filters.

I still would like to refactor the code to make the TMDB response be
more suited to be a unique Mailfin response.