Use Jellyfin data if TMDB ID is not provided
This provides a fallback mechanism to use Jellyfin data if the TMDB ID
is not provided in the request body. This is useful for when the user is
adding a movie or TV show that is not available on TMDB.

It seems to be the case that Jellyfin Webhhoks send a tvdb_id instead of
a TMDB ID for TV shows. This will allow the user to still receive an
email with the Jellyfin data if the TMDB ID is not provided.