Provide error message when TMDB ID is not found
This is needed as if the TMDB ID is not found, the user will not know
why the request failed. This will provide a more descriptive error
message. We check for an empty string because if the user uses a custom
handlebars it is possible that an empty string is passed as the TMDB ID.