Add better SMTP support for itemAdded route
This makes it so that we dynamically generate the subject line based on
the item type (Movie, Season, Episode) and include the name and release
year in the subject. The HTML template is also updated to better format
the email content. This helps to match it with what we have in
SendGrid.